ENH: add notebook to tokenize training data and push to S3 :sparkles:
The BlazingText algorithm requires the training and validation data to be in a sspecial version of the RecordIO format, so we conver the labeled solicitation documents from the Solicitation Review tool into that format and push to S3. This notebook assumes the user has already installed the awscli and configured it with their access id and key.